+++ lib/clerq/repositories/node_reader.rb
@@ -92,7 +92,7 @@ class Clerq::Repositories::NodeReader
     end
 
     def parse_meta(text)
-      text.strip.split(/[;\n]/).inject({}) do |h, i|
+      text.strip.split(/[;,\n]/).inject({}) do |h, i|
         pair = /\s?(\w*):\s*(.*)/.match(i)
         h.merge(pair[1].to_sym => pair[2])
       end || {}